moved non io methods
