+++ Assets/Lightspeed/Core/IO/FileHelper.cs
@@ -238,24 +238,6 @@ namespace Rhinox.Lightspeed.IO
                 return extension == null;
             return pathExt.Equals(extension, StringComparison.InvariantCultureIgnoreCase);
         }
-        
-        public static string GetLongestCommonPrefix(string[] s)
-        {
-            int k = s[0].Length;
-            for (int i = 1; i < s.Length; ++i)
-            {
-                k = Math.Min(k, s[i].Length);
-                for (int j = 0; j < k; j++)
-                {
-                    if (s[i][j] != s[0][j])
-                    {
-                        k = j;
-                        break;
-                    }
-                }
-            }
-            return s[0].Substring(0, k);
-        }
 
         public static string StripLastFolder(string s, bool separatorBackSlash = true, bool forceEndWithDirectorySeparator = false)
         {
@@ -272,11 +254,6 @@ namespace Rhinox.Lightspeed.IO
             if (forceEndWithDirectorySeparator)
                 result += targetSeparator;
             return result;
-        }
-
-        public static string[] SplitLines(this string input)
-        {
-            return input.Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None);
         }
     }
 }
